+++ ionoscloud/provider.go
@@ -71,7 +71,7 @@ func Provider() *schema.Provider {
 			ServerResource:             resourceServer(),
 			VolumeResource:             resourceVolume(),
 			GroupResource:              resourceGroup(),
-			"ionoscloud_share":         resourceShare(),
+			shareResource:              resourceShare(),
 			UserResource:               resourceUser(),
 			SnapshotResource:           resourceSnapshot(),
 			"ionoscloud_ipfailover":    resourceLanIPFailover(),
@@ -100,6 +100,7 @@ func Provider() *schema.Provider {
 			IpBLockResource:            dataSourceIpBlock(),
 			VolumeResource:             dataSourceVolume(),
 			nicResource:                dataSourceNIC(),
+			shareResource:              dataSourceShare(),
 		},
 	}
 	provider.ConfigureContextFunc = func(ctx context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {