+++ CHANGELOG.md
@@ -1,12 +1,12 @@
 ## 5.2.19
 
 - **code enhancements**: added http request time log for API calls
-- **code enhancements**: for `k8s_node_pool`:
+- **dependency update**: updated sdk-go to v5.1.9
+- **code enhancements**: for `k8s_node_pool`, `nic` and `share`:
   - made tests comprehensive
   - optimized test duration by including both match by id and by name in the same test
   - removed duplicated code from import, data_source, resource and tests files
-- **new features**: import for `nic`, data_source for `nic`
-- **dependency update**: updated sdk-go to v5.1.9
+- **new features**: import for `nic`, data_source for `nic`, `share`
 
 ## 5.2.18